+++ app/src/main/java/com/boostcamp/and03/ui/screen/bookdetail/BookDetailScreen.kt
@@ -48,8 +48,6 @@ import com.boostcamp.and03.ui.screen.bookdetail.component.DropdownMenuContainer
 import com.boostcamp.and03.ui.screen.bookdetail.component.MemoCard
 import com.boostcamp.and03.ui.component.QuoteCard
 import com.boostcamp.and03.ui.screen.bookdetail.component.SquareAddButton
-import com.boostcamp.and03.ui.screen.bookdetail.model.BookDetailAction
-import com.boostcamp.and03.ui.screen.bookdetail.model.BookDetailEvent
 import com.boostcamp.and03.ui.screen.bookdetail.model.BookDetailTab
 import com.boostcamp.and03.ui.screen.bookdetail.model.CharacterUiModel
 import com.boostcamp.and03.ui.screen.bookdetail.model.MemoType
