+++ app/src/main/java/com/boostcamp/and03/ui/screen/bookdetail/BookDetailViewModel.kt
@@ -7,8 +7,6 @@ import androidx.lifecycle.viewModelScope
 import androidx.navigation.toRoute
 import com.boostcamp.and03.data.repository.book_storage.BookStorageRepository
 import com.boostcamp.and03.ui.navigation.Route
-import com.boostcamp.and03.ui.screen.bookdetail.model.BookDetailAction
-import com.boostcamp.and03.ui.screen.bookdetail.model.BookDetailEvent
 import com.boostcamp.and03.ui.screen.bookdetail.model.toUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.collections.immutable.toPersistentList
