+++ app/src/main/java/com/boostcamp/and03/ui/screen/booksearch/BookSearchAction.kt
@@ -1,4 +1,6 @@
-package com.boostcamp.and03.ui.screen.booksearch.model
+package com.boostcamp.and03.ui.screen.booksearch
+
+import com.boostcamp.and03.ui.screen.booksearch.model.BookSearchResultUiModel
 
 sealed interface BookSearchAction {
