+++ app/src/main/java/com/boostcamp/and03/ui/screen/booksearch/BookSearchEvent.kt
@@ -1,4 +1,6 @@
-package com.boostcamp.and03.ui.screen.booksearch.model
+package com.boostcamp.and03.ui.screen.booksearch
+
+import com.boostcamp.and03.ui.screen.booksearch.model.SaveFailureReason
 
 sealed interface BookSearchEvent {
